+++ Sources/SwiftCIFP/Models/Airport/MSA.swift
@@ -122,7 +122,12 @@ extension MSA {
   public var centerFix: Fix? {
     get async {
       guard let data else { return nil }
-      return await data.resolveFix(center, sectionCode: nil, airportId: airportId)
+      return await data.resolveFix(
+        center,
+        icaoRegion: centerICAO,
+        sectionCode: nil,
+        airportId: airportId
+      )
     }
   }